# Add `@nx._dispatch` decorator to most algorithms