--- a/networkx/algorithms/approximation/clique.py
+++ b/networkx/algorithms/approximation/clique.py
@@ -11,9 +11,9 @@ __all__ = [
 ]
 
 
-@nx._dispatch
 @not_implemented_for("directed")
 @not_implemented_for("multigraph")
+@nx._dispatch
 def maximum_independent_set(G):
     """Returns an approximate maximum independent set.
 
@@ -68,9 +68,9 @@ def maximum_independent_set(G):
     return iset
 
 
-@nx._dispatch
 @not_implemented_for("directed")
 @not_implemented_for("multigraph")
+@nx._dispatch
 def max_clique(G):
     r"""Find the Maximum Clique
 
@@ -127,9 +127,9 @@ def max_clique(G):
     return iset
 
 
-@nx._dispatch
 @not_implemented_for("directed")
 @not_implemented_for("multigraph")
+@nx._dispatch
 def clique_removal(G):
     r"""Repeatedly remove cliques from the graph.
 
@@ -180,9 +180,9 @@ def clique_removal(G):
     return maxiset, cliques
 
 
-@nx._dispatch
 @not_implemented_for("directed")
 @not_implemented_for("multigraph")
+@nx._dispatch
 def large_clique_size(G):
     """Find the size of a large clique in a graph.