--- a/networkx/algorithms/approximation/clustering_coefficient.py
+++ b/networkx/algorithms/approximation/clustering_coefficient.py
@@ -1,8 +1,10 @@
+import networkx as nx
 from networkx.utils import not_implemented_for, py_random_state
 
 __all__ = ["average_clustering"]
 
 
+@nx._dispatch(name="approximate_average_clustering")
 @py_random_state(2)
 @not_implemented_for("directed")
 def average_clustering(G, trials=1000, seed=None):