--- a/networkx/algorithms/approximation/dominating_set.py
+++ b/networkx/algorithms/approximation/dominating_set.py
@@ -10,6 +10,7 @@ incident to an endpoint of at least one edge in *F*.
 .. _edge dominating set: https://en.wikipedia.org/wiki/Edge_dominating_set
 
 """
+import networkx as nx
 
 from ...utils import not_implemented_for
 from ..matching import maximal_matching
@@ -18,6 +19,7 @@ __all__ = ["min_weighted_dominating_set", "min_edge_dominating_set"]
 
 
 # TODO Why doesn't this algorithm work for directed graphs?
+@nx._dispatch(node_attrs="weight")
 @not_implemented_for("directed")
 def min_weighted_dominating_set(G, weight=None):
     r"""Returns a dominating set that approximates the minimum weight node
@@ -99,6 +101,7 @@ def min_weighted_dominating_set(G, weight=None):
     return dom_set
 
 
+@nx._dispatch
 def min_edge_dominating_set(G):
     r"""Returns minimum cardinality edge dominating set.