--- a/networkx/algorithms/approximation/kcomponents.py
+++ b/networkx/algorithms/approximation/kcomponents.py
@@ -13,6 +13,7 @@ from networkx.utils import not_implemented_for
 __all__ = ["k_components"]
 
 
+@nx._dispatch(name="approximate_k_components")
 @not_implemented_for("directed")
 def k_components(G, min_density=0.95):
     r"""Returns the approximate k-component structure of a graph G.