--- a/networkx/algorithms/approximation/matching.py
+++ b/networkx/algorithms/approximation/matching.py
@@ -13,6 +13,7 @@ import networkx as nx
 __all__ = ["min_maximal_matching"]
 
 
+@nx._dispatch
 def min_maximal_matching(G):
     r"""Returns the minimum maximal matching of G. That is, out of all maximal
     matchings of the graph G, the smallest is returned.