--- a/networkx/algorithms/approximation/traveling_salesman.py
+++ b/networkx/algorithms/approximation/traveling_salesman.py
@@ -123,6 +123,7 @@ def move_one_node(soln, seed):
     return soln
 
 
+@nx._dispatch(edge_attrs="weight")
 @not_implemented_for("directed")
 def christofides(G, weight="weight", tree=None):
     """Approximate a solution of the traveling salesman problem
@@ -196,6 +197,7 @@ def _shortcutting(circuit):
     return nodes
 
 
+@nx._dispatch(edge_attrs="weight")
 def traveling_salesman_problem(G, weight="weight", nodes=None, cycle=True, method=None):
     """Find the shortest path in `G` connecting specified nodes
 
@@ -333,6 +335,7 @@ def traveling_salesman_problem(G, weight="weight", nodes=None, cycle=True, metho
     return best_path
 
 
+@nx._dispatch(edge_attrs="weight")
 @not_implemented_for("undirected")
 @py_random_state(2)
 def asadpour_atsp(G, weight="weight", seed=None, source=None):
@@ -484,6 +487,7 @@ def asadpour_atsp(G, weight="weight", seed=None, source=None):
     return _shortcutting(circuit)
 
 
+@nx._dispatch(edge_attrs="weight")
 def held_karp_ascent(G, weight="weight"):
     """
     Minimizes the Held-Karp relaxation of the TSP for `G`
@@ -794,6 +798,7 @@ def held_karp_ascent(G, weight="weight"):
     return next(k_max.__iter__()).size(weight), z_star
 
 
+@nx._dispatch
 def spanning_tree_distribution(G, z):
     """
     Find the asadpour exponential distribution of spanning trees.
@@ -904,6 +909,7 @@ def spanning_tree_distribution(G, z):
     return gamma
 
 
+@nx._dispatch(edge_attrs="weight")
 def greedy_tsp(G, weight="weight", source=None):
     """Return a low cost cycle starting at `source` and its cost.
 
@@ -995,6 +1001,7 @@ def greedy_tsp(G, weight="weight", source=None):
     return cycle
 
 
+@nx._dispatch(edge_attrs="weight")
 @py_random_state(9)
 def simulated_annealing_tsp(
     G,
@@ -1213,6 +1220,7 @@ def simulated_annealing_tsp(
     return best_cycle
 
 
+@nx._dispatch(edge_attrs="weight")
 @py_random_state(9)
 def threshold_accepting_tsp(
     G,