--- a/networkx/algorithms/approximation/vertex_cover.py
+++ b/networkx/algorithms/approximation/vertex_cover.py
@@ -7,10 +7,12 @@ is incident to at least one node in the subset.
 .. |vertex cover| replace:: *vertex cover*
 
 """
+import networkx as nx
 
 __all__ = ["min_weighted_vertex_cover"]
 
 
+@nx._dispatch(node_attrs="weight")
 def min_weighted_vertex_cover(G, weight=None):
     r"""Returns an approximate minimum weighted vertex cover.