--- a/networkx/algorithms/assortativity/correlation.py
+++ b/networkx/algorithms/assortativity/correlation.py
@@ -1,5 +1,6 @@
 """Node assortativity coefficients and correlation measures.
 """
+import networkx as nx
 from networkx.algorithms.assortativity.mixing import (
     attribute_mixing_matrix,
     degree_mixing_matrix,
@@ -14,6 +15,7 @@ __all__ = [
 ]
 
 
+@nx._dispatch(edge_attrs="weight")
 def degree_assortativity_coefficient(G, x="out", y="in", weight=None, nodes=None):
     """Compute degree assortativity of graph.
 
@@ -98,6 +100,7 @@ def degree_assortativity_coefficient(G, x="out", y="in", weight=None, nodes=None
     return _numeric_ac(M, mapping=mapping)
 
 
+@nx._dispatch(edge_attrs="weight")
 def degree_pearson_correlation_coefficient(G, x="out", y="in", weight=None, nodes=None):
     """Compute degree assortativity of graph.
 
@@ -157,6 +160,7 @@ def degree_pearson_correlation_coefficient(G, x="out", y="in", weight=None, node
     return sp.stats.pearsonr(x, y)[0]
 
 
+@nx._dispatch(node_attrs="attribute")
 def attribute_assortativity_coefficient(G, attribute, nodes=None):
     """Compute assortativity for node attributes.
 
@@ -203,6 +207,7 @@ def attribute_assortativity_coefficient(G, attribute, nodes=None):
     return attribute_ac(M)
 
 
+@nx._dispatch(node_attrs="attribute")
 def numeric_assortativity_coefficient(G, attribute, nodes=None):
     """Compute assortativity for numerical node attributes.