--- a/networkx/algorithms/assortativity/pairs.py
+++ b/networkx/algorithms/assortativity/pairs.py
@@ -1,7 +1,10 @@
 """Generators of  x-y pairs of node data."""
+import networkx as nx
+
 __all__ = ["node_attribute_xy", "node_degree_xy"]
 
 
+@nx._dispatch(node_attrs="attribute")
 def node_attribute_xy(G, attribute, nodes=None):
     """Returns iterator of node-attribute pairs for all edges in G.
 
@@ -56,6 +59,7 @@ def node_attribute_xy(G, attribute, nodes=None):
                 yield (uattr, vattr)
 
 
+@nx._dispatch(edge_attrs="weight")
 def node_degree_xy(G, x="out", y="in", weight=None, nodes=None):
     """Generate node degree-degree pairs for edges in G.