--- a/networkx/algorithms/asteroidal.py
+++ b/networkx/algorithms/asteroidal.py
@@ -16,6 +16,7 @@ from networkx.utils import not_implemented_for
 __all__ = ["is_at_free", "find_asteroidal_triple"]
 
 
+@nx._dispatch
 @not_implemented_for("directed")
 @not_implemented_for("multigraph")
 def find_asteroidal_triple(G):
@@ -88,6 +89,7 @@ def find_asteroidal_triple(G):
     return None
 
 
+@nx._dispatch
 @not_implemented_for("directed")
 @not_implemented_for("multigraph")
 def is_at_free(G):
@@ -121,6 +123,7 @@ def is_at_free(G):
     return find_asteroidal_triple(G) is None
 
 
+@nx._dispatch
 @not_implemented_for("directed")
 @not_implemented_for("multigraph")
 def create_component_structure(G):