--- a/networkx/algorithms/bipartite/basic.py
+++ b/networkx/algorithms/bipartite/basic.py
@@ -17,6 +17,7 @@ __all__ = [
 ]
 
 
+@nx._dispatch
 def color(G):
     """Returns a two-coloring of the graph.
 
@@ -108,6 +109,7 @@ def is_bipartite(G):
         return False
 
 
+@nx._dispatch
 def is_bipartite_node_set(G, nodes):
     """Returns True if nodes and G/nodes are a bipartition of G.
 
@@ -152,6 +154,7 @@ def is_bipartite_node_set(G, nodes):
     return True
 
 
+@nx._dispatch
 def sets(G, top_nodes=None):
     """Returns bipartite node sets of graph G.
 
@@ -218,6 +221,7 @@ def sets(G, top_nodes=None):
     return (X, Y)
 
 
+@nx._dispatch(graphs="B")
 def density(B, nodes):
     """Returns density of bipartite graph B.
 
@@ -270,6 +274,7 @@ def density(B, nodes):
     return d
 
 
+@nx._dispatch(graphs="B", edge_attrs="weight")
 def degrees(B, nodes, weight=None):
     """Returns the degrees of the two node sets in the bipartite graph B.