--- a/networkx/algorithms/bipartite/covering.py
+++ b/networkx/algorithms/bipartite/covering.py
@@ -1,5 +1,6 @@
 """ Functions related to graph covers."""
 
+import networkx as nx
 from networkx.algorithms.bipartite.matching import hopcroft_karp_matching
 from networkx.algorithms.covering import min_edge_cover as _min_edge_cover
 from networkx.utils import not_implemented_for
@@ -7,6 +8,7 @@ from networkx.utils import not_implemented_for
 __all__ = ["min_edge_cover"]
 
 
+@nx._dispatch(name="bipartite_min_edge_cover")
 @not_implemented_for("directed")
 @not_implemented_for("multigraph")
 def min_edge_cover(G, matching_algorithm=None):