--- a/networkx/algorithms/bipartite/covering.py
+++ b/networkx/algorithms/bipartite/covering.py
@@ -8,9 +8,9 @@ from networkx.utils import not_implemented_for
 __all__ = ["min_edge_cover"]
 
 
-@nx._dispatch(name="bipartite_min_edge_cover")
 @not_implemented_for("directed")
 @not_implemented_for("multigraph")
+@nx._dispatch(name="bipartite_min_edge_cover")
 def min_edge_cover(G, matching_algorithm=None):
     """Returns a set of edges which constitutes
     the minimum edge cover of the graph.