--- a/networkx/algorithms/bipartite/projection.py
+++ b/networkx/algorithms/bipartite/projection.py
@@ -12,6 +12,7 @@ __all__ = [
 ]
 
 
+@nx._dispatch(graphs="B")
 def projected_graph(B, nodes, multigraph=False):
     r"""Returns the projection of B onto one of its node sets.
 
@@ -115,6 +116,7 @@ def projected_graph(B, nodes, multigraph=False):
     return G
 
 
+@nx._dispatch(graphs="B")
 @not_implemented_for("multigraph")
 def weighted_projected_graph(B, nodes, ratio=False):
     r"""Returns a weighted projection of B onto one of its node sets.
@@ -215,6 +217,7 @@ def weighted_projected_graph(B, nodes, ratio=False):
     return G
 
 
+@nx._dispatch(graphs="B")
 @not_implemented_for("multigraph")
 def collaboration_weighted_projected_graph(B, nodes):
     r"""Newman's weighted projection of B onto one of its node sets.
@@ -309,6 +312,7 @@ def collaboration_weighted_projected_graph(B, nodes):
     return G
 
 
+@nx._dispatch(graphs="B")
 @not_implemented_for("multigraph")
 def overlap_weighted_projected_graph(B, nodes, jaccard=True):
     r"""Overlap weighted projection of B onto one of its node sets.
@@ -408,6 +412,7 @@ def overlap_weighted_projected_graph(B, nodes, jaccard=True):
     return G
 
 
+@nx._dispatch(graphs="B")
 @not_implemented_for("multigraph")
 def generic_weighted_projected_graph(B, nodes, weight_function=None):
     r"""Weighted projection of B with a user-specified weight function.