--- a/networkx/algorithms/bipartite/redundancy.py
+++ b/networkx/algorithms/bipartite/redundancy.py
@@ -1,11 +1,13 @@
 """Node redundancy for bipartite graphs."""
 from itertools import combinations
 
+import networkx as nx
 from networkx import NetworkXError
 
 __all__ = ["node_redundancy"]
 
 
+@nx._dispatch
 def node_redundancy(G, nodes=None):
     r"""Computes the node redundancy coefficients for the nodes in the bipartite
     graph `G`.