This adds dispatching to most algorithms (only functions in `algorithms`, `generators`, and `linalg`). This raises a few questions that we need to solve or postpone. I will raise these questions below.

CC @jim22k @MridulS