@eriknw 2023-05-13
This brings the total number of dispatched functions up to 544. We currently implement 81 in (my local branch of) `graphblas-algorithms`, which is about 15%.