@eriknw 2023-06-01
I updated `@nx._dispatch` based on feedback from @jim22k. We still need to update documentation and make sure everything works.

A known shortcoming is handling of functions that return Graphs--we don't always indicate whether edge, node, and/or graph attributes should be preserved. Also, we don't convert graph attributes that are _graphs_. If a reviewer/contributor wants to take a closer look at this, it would be appreciated. Otherwise, I'm inclined to handle these in the future when needed.

Also, I haven't benchmarked/profiled impact to import time or dispatched function call overhead.