@eriknw 2023-06-05
@MridulS I think this is ready for review. Who else can/should review this? Do you want this PR split into two: one that adds the `@nx._dispatch` decorator everywhere, and another that updates the `def _dispatch` function? I don't have a strong opinion on this, and I think a single monolithic PR is fine.

Note that major changes were made to `networkx/utils/backends.py` that defines the `_dispatch` decorator. This file used to be called `networkx/classes/backends.py`, but was moved to avoid difficulty importing.

I have [`graphblas-algorithms`](https://github.com/python-graphblas/graphblas-algorithms) working against this PR.