@eriknw 2023-06-07
> Do you know if the mypy complain has anything to do with this PR?

I have no idea what mypy complaining about.