@eriknw 2023-06-17
My latest changes are up, and I don't anticipate making any more major changes.

Actually testing the values of `@nx._dispatch` arguments uncovered a bunch of things, including a couple docstring errors.

So, yeah, the `nx-loopback` "backend" now converts input graphs to only include the specified edge, node, and graph attributes. This give us _much_ greater confidence in the correctness of the dispatch metadata.  When adding `preserve_edge_attrs=`, `preserve_node_attrs=`, and `preserve_graph_attrs=`, I heavily relied on the tests, not the docstrings. IMHO, this means that the most likely mistakes are algorithms that _should_ preserve edge, node, or graph attributes, but aren't annotated as such because the behavior isn't exercised in tests. I don't think this should worry us too much--each function should get closer scrutiny when implemented by backends.

As a "first pass" to dispatch all the things, this PR is much more functional and complete than I anticipated. It took a lot of work. I think it lays a solid foundation for being able to automatically convert to different backends.

fwiw, I actually like seeing metadata in `@nx._dispatch` about the "shape" of the inputs needed for the calculation. We don't yet have similar metadata describing output graphs or how input graphs are mutated. In the future, I think it would also be handy to describe how multigraph attributes could be combined (add, first, last, etc.) to be turned into a (not multi-) graph for algorithms.