@eriknw 2023-06-25
I wonder whether we should have `@nx._dispatch` decorator _after_ `@not_implemented_for` decorator. Even if the input graph isn't from networkx, it ought to be easy for it to implement `g.is_multigraph()` and `g.is_directed()`.

Once we enable conversions to other backends, it seems silly to perform the conversion merely to have it raise b/c it's the wrong type of graph. On the other hand, I don't really understand how this may interact with `argmap` (used by `not_implemented_for`--will it still work if `@nx._dispatch` returns a class instance and not a normal Python function?) and `@py_random_state`.