@eriknw 2023-06-28
This PR doesn't convert graphs used as edge attributes. For example, this happens in `test_christofides_hamiltonian`.