@dschult 2023-07-16
Other notes:
- The default value for `partition` in the `tree/branching.py` functions is `0` or if you prefer you can use `nx.EdgePartition.OPEN`
- I do not find a function to reduce a weighted multiedge set for two nodes to a single edge weight for those two nodes. There is an idiom used in a few examples though:
```python
G=nx.Graph(MG)
for u, v in MG.edges():   # parens edges() are needed to suppress 3-tuple edge reporting
    G[u][v]["weight"] = min(d.get("weight", 1) for d in MG[u][v].values())
```
where `min` can be replaced by `sum`, `max`, `np.mean` and other reduction functions.