@eriknw 2023-07-16
Alright, I believe I'm made all the requested updates. Thanks for the careful review @dschult! I think the last remaining thing is fixing the Mypy issue.

Possible follow-up work:
- Measure impact of dispatching on import time and function call time
- Use a class for dispatch function instead of a closure
- Have dispatch decorator indicate whether multigraph edges can be merged with e.g. `sum` or `min`
- Add `@nx._dispatch` to appropriate functions in `networkx.drawing` (anywhere else?)