@dschult 2023-07-22
Here are some crude timing results:
    Full test suite for networkx: (very crude test using pytest time reporting)
      - without dispatch decorators (before PR #6688): 48.14s   
      - with dispatch decorators: 48.83s
      - in nx-loopback test mode without dispatch decorators (should take longer due to converting graphs): 56.37s
      - in nx-loopback test mode without dispatch decorators (should take longer due to converting graphs): 154.24s

    Time a function call that takes very little time: `G=nx.path_graph(5); %timeit nx.shortest_path(0,3)`
      - without dispatch decorators: 2.67 µs
      - with dispatch decorators: 3.3 µs
    Increase number of nodes:  `G=nx.path_graph(50);%timeit nx.shortest_path(G, 0, 47)`
      - without dispatch decorators: 21.9 µs
      - with dispatch decorators: 22.3 µs

Timing of import of networkx:   
    Change to new git branch, then `python -X importtime -c "import networkx"`
      - without dispatch decorators: ~70 +/-2 ms    (200+/-4 ms for first invocation after switching branches)
      - with dispatch decorators: ~72 +/-2 ms    (204+/-4 ms for first invocation after switching branches)
The first time an import occurs, a py_cache is stored which speeds things up for subsequent imports.