@eriknw 2023-07-26
Idea: add `preserve_all_attrs=True` argument to `_dispatch` and `convert_from_nx` that is equivalent to `preserve_edge_attrs=True, preserve_node_attrs=True, preserve_graph_attrs=True`.