@dschult 2023-07-27
This sounds like a good idea to me. It will cover what seems like a common case and make the code easier to read too.